Cryptography is done

+++ src/Cryptography/Cryptography.Core/Repositories/IPaymentRepository.cs
@@ -5,7 +5,7 @@ namespace Cryptography.Repositories;
 public interface IPaymentRepository
 {
     Task<IList<Payment>> GetAllAsync();
-    Task<Payment> FindAsync(int id);
+    Task<Payment> FindAsync(long id);
     Task CreateAsync(Payment entity);
     Task UpdateAsync(Payment entity);
     Task RemoveAsync(long id);
